+++ db/access.php
@@ -17,7 +17,7 @@
  * Capabilities for stripe enrolment plugin.
  *
  * @package    enrol_stripepayment
- * @copyright  2019 Dualcube Team
+ * @copyright  2023 DualCube Team
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 defined('MOODLE_INTERNAL') || die();