+++ edit_form.php
@@ -16,42 +16,39 @@
 /**
  * Course wise edit form.
  *
- * Adds new instance of enrol_stripepayment to specified course
- * or edits current instance.
+ * This plugin allows you to set up paid courses.
  *
  * @package    enrol_stripepayment
- * @copyright  2019 Dualcube Team
+ * @author     DualCube <admin@dualcube.com>
+ * @copyright  DualCube (https://dualcube.com)
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
 defined('MOODLE_INTERNAL') || die();
-require_once($CFG->libdir . '/formslib.php');
+require_once($CFG->libdir.'/formslib.php');
 require_once('lib.php');
 /**
  * Sets up moodle edit form class methods.
- * @copyright  2019 Dualcube Team
+ * @copyright  2023 DualCube Team
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class enrol_stripepayment_edit_form extends moodleform
-{
+class enrol_stripepayment_edit_form extends moodleform {
     /**
      * Sets up moodle form.
      * @return void
      */
-    public function definition()
-    {
+    public function definition() {
         $mform = $this->_form;
         list($instance, $plugin, $context) = $this->_customdata;
         $mform->addElement('header', 'header', get_string('pluginname', 'enrol_stripepayment'));
         $mform->addElement('text', 'name', get_string('custominstancename', 'enrol'));
         $mform->setType('name', PARAM_TEXT);
-        $options = array(
-            ENROL_INSTANCE_ENABLED  => get_string('yes'),
-            ENROL_INSTANCE_DISABLED => get_string('no')
-        );
+        $options = array(ENROL_INSTANCE_ENABLED  => get_string('yes'),
+                         ENROL_INSTANCE_DISABLED => get_string('no'));
         $mform->addElement('select', 'status', get_string('status', 'enrol_stripepayment'), $options);
         $mform->setDefault('status', $plugin->get_config('status'));
         $costarray = array();
-        $costarray[] = &$mform->createElement('text', 'cost', get_string('cost', 'enrol_stripepayment'), array('size' => 4));
+        $costarray[] =& $mform->createElement('text', 'cost', get_string('cost', 'enrol_stripepayment'), array('size' => 4));
         $mform->setDefault('cost', format_float($plugin->get_config('cost'), 2, true));
         $mform->setType('cost', PARAM_INT);
         $mform->addGroup($costarray, 'costar', get_string('cost', 'enrol_stripepayment'), array(' '), false);
@@ -71,28 +68,16 @@ class enrol_stripepayment_edit_form extends moodleform
         $mform->setDefault('maxenrolled', 'customint3');
         $mform->addHelpButton('customint3', 'maxenrolled', 'enrol_stripepayment');
         $mform->setType('customint3', PARAM_INT);
-        $mform->addElement(
-            'duration',
-            'enrolperiod',
-            get_string('enrolperiod', 'enrol_stripepayment'),
-            array('optional' => true, 'defaultunit' => 86400)
-        );
+        $mform->addElement('duration', 'enrolperiod', get_string('enrolperiod', 'enrol_stripepayment'),
+        array('optional' => true, 'defaultunit' => 86400));
         $mform->setDefault('enrolperiod', $plugin->get_config('enrolperiod'));
         $mform->addHelpButton('enrolperiod', 'enrolperiod', 'enrol_stripepayment');
-        $mform->addElement(
-            'date_time_selector',
-            'enrolstartdate',
-            get_string('enrolstartdate', 'enrol_stripepayment'),
-            array('optional' => true)
-        );
+        $mform->addElement('date_time_selector', 'enrolstartdate', get_string('enrolstartdate', 'enrol_stripepayment'),
+        array('optional' => true));
         $mform->setDefault('enrolstartdate', 0);
         $mform->addHelpButton('enrolstartdate', 'enrolstartdate', 'enrol_stripepayment');
-        $mform->addElement(
-            'date_time_selector',
-            'enrolenddate',
-            get_string('enrolenddate', 'enrol_stripepayment'),
-            array('optional' => true)
-        );
+        $mform->addElement('date_time_selector', 'enrolenddate', get_string('enrolenddate', 'enrol_stripepayment'),
+        array('optional' => true));
         $mform->setDefault('enrolenddate', 0);
         $mform->addHelpButton('enrolenddate', 'enrolenddate', 'enrol_stripepayment');
         $mform->addElement('hidden', 'id');
@@ -100,12 +85,8 @@ class enrol_stripepayment_edit_form extends moodleform
         $mform->addElement('hidden', 'courseid');
         $mform->setType('courseid', PARAM_INT);
         if (enrol_accessing_via_instance($instance)) {
-            $mform->addElement(
-                'static',
-                'selfwarn',
-                get_string('instanceeditselfwarning', 'core_enrol'),
-                get_string('instanceeditselfwarningtext', 'core_enrol')
-            );
+            $mform->addElement('static', 'selfwarn', get_string('instanceeditselfwarning', 'core_enrol'),
+            get_string('instanceeditselfwarningtext', 'core_enrol'));
         }
         $this->add_action_buttons(true, ($instance->id ? null : get_string('addinstance', 'enrol')));
         $this->set_data($instance);
@@ -116,8 +97,7 @@ class enrol_stripepayment_edit_form extends moodleform
      * @param stdClass $files
      * @return $error error list
      */
-    public function validation($data, $files)
-    {
+    public function validation($data, $files) {
         global $DB, $CFG;
         $errors = parent::validation($data, $files);
         list($instance, $plugin, $context) = $this->_customdata;