+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  * Fixed Chackout error.
  * Fixed Expired Coupon Faild silently.
  * FIxed Payment description no stripe dashboard not shows.
+ * Fixed coupon apply resister in Stripe.
  * Fixed internal error.
 
 3.3.4 (2022070702)